Data/requirements.txt
Pillow==11.3.0
beautifulsoup4==4.14.3
deep-translator==1.11.4
python-dotenv==1.2.1
langdetect==1.0.9
anthropic==0.97.0
requests==2.32.5
mysql-connector-python==9.4.0
numpy==2.0.2
